Fix build errors: duplicate ChildFeedback, clamp() type casts, missing theme import, fl_chart x cast

+++ lib/screens/profile_screen.dart
@@ -414,7 +414,7 @@ class _ProfileScreenState extends State<ProfileScreen> {
                       .asMap()
                       .entries
                       .map((e) => BarChartGroupData(
-                            x: e.key,
+                            x: e.key.toDouble(),
                             barRods: [
                               BarChartRodData(
                                 toY: e.value.value.toDouble(),
